# specs/001-build-a-comprehensive/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Amharic Document Preparation System API
  description: |
    Comprehensive API for processing, analyzing, and managing Amharic documents with multi-agent orchestration.
    
    ## Features
    - Multi-format document upload (PDF, images, Word, CSV, web content)
    - Real-time processing with CrewAI agents
    - Amharic-specific text analysis and OCR
    - Advanced search and export capabilities
    - Quality assurance and validation metrics
    
  version: 1.0.0
  contact:
    name: Amharic Document System Team
    email: support@amharic-docs.ai
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.amharic-docs.ai/v1
    description: Production server
...
paths:
...
  /documents/{document_id}:
    get:
      tags: [Documents]
      summary: Get document details
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Document not found

    delete:
      tags: [Documents]
      summary: Delete document
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found
...
  /export/documents/{document_id}:
    post:
      tags: [Export]
      summary: Export document
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Document exported successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Document not found
  /export/templates:
    get:
      tags: [Export]
      summary: List export templates
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Export templates retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExportTemplate'

  # Quality Metrics Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        filename:
          type: string
        document_type:
          $ref: '#/components/schemas/DocumentType'
        file_size:
          type: integer
        content_hash:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        metadata:
          $ref: '#/components/schemas/DocumentMetadata'
        created_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
    DocumentType:
      type: string
      enum: [pdf, image, word, csv, web_content]
    DocumentStatus:
      type: string
      enum: [uploaded, queued, processing, completed, failed, archived]
    DocumentMetadata:
      type: object
      properties:
        original_name:
          type: string
        mime_type:
          type: string
        page_count:
          type: integer
        language_detected:
          type: string
        creation_date:
          type: string
          format: date-time
        author:
          type: string
...
    ExportRequest:
      type: object
      required:
        - format
      properties:
        format:
          type: string
          enum: [pdf, docx, html, markdown, json]
        template_id:
          type: string
          format: uuid
        include_metadata:
          type: boolean
          default: true
        include_quality_metrics:
          type: boolean
          default: false
    ExportTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        output_format:
          type: string
          enum: [pdf, docx, html, markdown, json]
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time

    # Quality Schemas
...